+++ rmf_traffic/CHANGELOG.rst
@@ -5,6 +5,11 @@ Changelog for package rmf_traffic
 Forthcoming
 -----------
 
+1.0.1 (2020-07-XX)
+------------------
+* Allow users to specify a callback for interrupting a planner: [#130](https://github.com/osrf/rmf_core/pull/130/)
+* Allow a Negotiation Table Viewer to know when its Table is defunct: [#130](https://github.com/osrf/rmf_core/pull/130/)
+
 1.0.0 (2020-06-23)
 ------------------
 * Provides core `rmf_traffic` utilities